Moved reduxform from QuoteEditorRoute to AddQuote and fixed tests
