+++ app/routes/quotes/components/__tests__/AddQuote.spec.js
@@ -1,27 +1,25 @@
 import React from 'react';
 import AddQuote from '../AddQuote';
 import { mount } from 'enzyme';
+import configureStore from 'app/utils/configureStore';
+import { Provider } from 'react-redux';
 
-const emptyProps = () => ({
-  // Just to remove warnings from npm test
-  fields: {
-    text: {},
-    source: {}
-  },
-  invalid: true,
-  pristine: true,
-  submitting: false
-});
 
 describe('components', () => {
   describe('AddQuote', () => {
     it('should be possible to submit forms', () => {
       const addQuotes = jest.fn();
+      const validator = jest.fn(() => ({}));
+      const store = configureStore();
       const wrapper = mount(
-        <AddQuote addQuotes={addQuotes} {...emptyProps()} />
+        <Provider {...{ store }}>
+          <AddQuote addQuotes={addQuotes} validate={validator} />
+        </Provider>
       );
       const form = wrapper.find('form');
       form.simulate('submit');
+
+      expect(validator).toHaveBeenCalled();
       expect(addQuotes).toHaveBeenCalled();
     });
   });
